# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/lessons/{id}/progress": {
      "post": {
        "operationId": "LessonsController_progress",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LessonProgressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "ok": true
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Lessons"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/lessons/{id}/complete": {
      "post": {
        "operationId": "LessonsController_complete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "ok": true,
                    "progressPercentage": 100,
                    "isCompleted": true
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Lessons"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "KODIRA API",
    "description": "",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "LessonProgressDto": {
        "type": "object",
        "properties": {
          "watchPercentage": {
            "type": "number",
            "example": 25
          },
          "lastPositionSeconds": {
            "type": "number",
            "example": 90
          },
          "quizScore": {
            "type": "number",
            "nullable": true
          }
        },
        "required": [
          "watchPercentage",
          "lastPositionSeconds"
        ]
      },
...
    }
  }
}